requirements.txt
python-dotenv == 1.1.1
mailchimp_marketing == 3.0.80
boto3 == 1.39.7